Extract common imgur uploading from Submit and DoEditorActions
