+++ app/src/main/java/me/ccrama/redditslide/util/ProgressRequestBody.java
@@ -1,0 +1,66 @@
+package me.ccrama.redditslide.util;
+
+import java.io.IOException;
+
+import okhttp3.MediaType;
+import okhttp3.RequestBody;
+import okio.Buffer;
+import okio.BufferedSink;
+import okio.ForwardingSink;
+import okio.Okio;
+import okio.Sink;
+
+// Used in DoEditorActions and Submit
+public class ProgressRequestBody extends RequestBody {
+
+    protected RequestBody mDelegate;
+    protected Listener mListener;
+    protected CountingSink mCountingSink;
+
+    public ProgressRequestBody(RequestBody delegate, Listener listener) {
+        mDelegate = delegate;
+        mListener = listener;
+    }
+
+    @Override
+    public MediaType contentType() {
+        return mDelegate.contentType();
+    }
+
+    @Override
+    public long contentLength() {
+        try {
+            return mDelegate.contentLength();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return -1;
+    }
+
+    @Override
+    public void writeTo(BufferedSink sink) throws IOException {
+        mCountingSink = new CountingSink(sink);
+        BufferedSink bufferedSink = Okio.buffer(mCountingSink);
+        mDelegate.writeTo(bufferedSink);
+        bufferedSink.flush();
+    }
+
+    public interface Listener {
+        void onProgress(int progress);
+    }
+
+    protected final class CountingSink extends ForwardingSink {
+        private long bytesWritten = 0;
+
+        public CountingSink(Sink delegate) {
+            super(delegate);
+        }
+
+        @Override
+        public void write(Buffer source, long byteCount) throws IOException {
+            super.write(source, byteCount);
+            bytesWritten += byteCount;
+            mListener.onProgress((int) (100F * bytesWritten / contentLength()));
+        }
+    }
+}
